Moved helper functions from FormRow to separate file in order to solve problems where FormRows could not be overriden
Added missing import in index.js

Change back to latest component-registry syntax
