+++ src/widgets/FormRow.js
@@ -22,63 +22,7 @@ import {
     Label as _bs_Label
  } from 'inferno-bootstrap'
 
-function Label (props) {
-    return <_bs_Label>{renderString(props.children, props.options && props.options.lang)}</_bs_Label>
-}
-
-function HelpMsg (props, context) {
-    // don't render if nothing to show
-    if (!props.text && !props.required) return null
-
-    const outp = []
-    if (props.text) outp.push(renderString(props.text, props.options && props.options.lang))
-    if (props.required) outp.push(renderString(i18n('isomorphic-schema--field_required', '(required)'), props.options && props.options.lang, '(required)'))
-
-    if (context && context.renderHelpAsHtml) {
-        return <FormText className="text-muted" for={props.id} dangerouslySetInnerHTML={{ __html: outp.join(' ')}} />
-    } else {
-        return <FormText className="text-muted" for={props.id}>{outp.join(' ')}</FormText>
-    }
-}
-
-class ErrorMsg extends Component {
-
-    componentDidMount () {
-        animateOnAdd(findDOMNode(this), 'InfernoFormlib-ErrorMsg--Animation')
-    }
-
-    componentWillUnmount () {
-        animateOnRemove(findDOMNode(this), 'InfernoFormlib-ErrorMsg--Animation')
-    }
-
-    render () {
-        return <FormFeedback>{renderString(this.props.validationError.i18nLabel, this.props.options && this.props.options.lang, this.props.validationError.message)}</FormFeedback>
-    }
-}
-
-function unpackInvariantErrors (validationError, namespace) {
-    if (validationError === undefined) return
-
-    let invariantError
-    if (Array.isArray(validationError.invariantErrors)){
-        
-        const dotName = namespace.join('.')
-        let tmpInvErrs = validationError.invariantErrors.filter((invErr) => {
-            // Pattern match field name of error with current namespace to see if it is a match
-            return invErr.fields.reduce((prev, curr) => prev || (curr === dotName), false)
-        })
-        
-        if (tmpInvErrs.length > 1) {
-            invariantError = {
-                message: 'There are several form value errors here',
-                i18nLabel: 'inferno-formlib--multiple_invariant_errors'
-            }
-        } else {
-            invariantError = tmpInvErrs[0]
-        }
-    }
-    return invariantError
-}
+import { ErrorMsg, HelpMsg, unpackInvariantErrors, Label} from './validation'
 
 /*
     PROPS:
